requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
pyTelegramBotAPI==4.2.2
python-decouple==3.6
requests==2.27.0
urllib3==1.26.7